// src/redux/actions/placesActions/types.ts
export enum Actions {
  GET_ALL_PLACES_SUCCESS = 'GET_ALL_PLACES_FAIL',
  GET_ALL_PLACES_FAIL = 'GET_ALL_PLACES_FAIL',
  CREATE_PLACE_SUCCESS = 'CREATE_PLACE_SUCCESS',
  CREATE_PLACE_FAIL = 'CREATE_PLACE_FAIL',
  DELETE_PLACE_SUCCESS = 'DELETE_PLACE_SUCCESS',
  DELETE_PLACE_FAIL = 'DELETE_PLACE_FAIL',
  UPDATE_PLACE_SUCCESS = 'UPDATE_PLACE_SUCCESS',
  UPDATE_PLACE_FAIL = 'UPDATE_PLACE_FAIL',
}